调整会员中心部分接口入参类型

+++ src/main/java/com/zpl/eshop/membership/service/impl/MembershipServiceImpl.java
@@ -47,7 +47,7 @@ public class MembershipServiceImpl implements MembershipService {
      * @return 处理结果
      */
     @Override
-    public Boolean informPayOrderEvent(Long userAccountId, Long totalOrderAmount) {
+    public Boolean informPayOrderEvent(Long userAccountId, Double totalOrderAmount) {
         return true;
     }
 
@@ -59,7 +59,7 @@ public class MembershipServiceImpl implements MembershipService {
      * @return 处理结果
      */
     @Override
-    public Boolean informFinishReturnGoodsEvent(Long userAccountId, Long totalOrderAmount) {
+    public Boolean informFinishReturnGoodsEvent(Long userAccountId, Double totalOrderAmount) {
         return true;
     }
